Renamed taxonomyio directory
To match the package name.
