+++ taxonomyio/write.go
